refactor - update file organization of section components
