--- src/ui/section/components/PlaybackActionArea.h
@@ -1,77 +1,0 @@
-#pragma once
-
-#include <JuceHeader.h>
-#include "ui/extra/ColoredPanel.h"
-#include "ui/extra/StyledButton.h"
-
-//==============================================================================
-class PlaybackActionArea : public ColoredPanel
-{
-public:
-    PlaybackActionArea() 
-        : ColoredPanel(juce::Colour::fromString("#ffededfc")),
-          playButton(juce::String::fromUTF8("▶"), juce::Colours::black, juce::Colour::fromString("#ff666666"), 16.0f, FontManager::FontWeight::Bold),
-          pauseButton(juce::String::fromUTF8("⏸"), juce::Colours::black, juce::Colour::fromString("#ff666666"), 16.0f, FontManager::FontWeight::Bold),
-          stopButton(juce::String::fromUTF8("⏹"), juce::Colours::black, juce::Colour::fromString("#ff666666"), 16.0f, FontManager::FontWeight::Bold)
-    {
-        // Ajouter les boutons
-        addAndMakeVisible(playButton);
-        addAndMakeVisible(pauseButton);
-        addAndMakeVisible(stopButton);
-        
-        // Configuration du FlexBox
-        flexBox.flexDirection = juce::FlexBox::Direction::row;
-        flexBox.justifyContent = juce::FlexBox::JustifyContent::center;
-        flexBox.alignItems = juce::FlexBox::AlignItems::center;
-        flexBox.flexWrap = juce::FlexBox::Wrap::noWrap;
-        
-        // Ajouter les boutons au FlexBox avec espacement
-        flexBox.items.add(juce::FlexItem(playButton).withWidth(32.0f).withHeight(32.0f).withMargin(juce::FlexItem::Margin(0, 4, 0, 0)));
-        flexBox.items.add(juce::FlexItem(pauseButton).withWidth(32.0f).withHeight(32.0f).withMargin(juce::FlexItem::Margin(0, 4, 0, 0)));
-        flexBox.items.add(juce::FlexItem(stopButton).withWidth(32.0f).withHeight(32.0f).withMargin(juce::FlexItem::Margin(0, 0, 0, 0)));
-    }
-    
-    void paint(juce::Graphics& g) override
-    {
-        ColoredPanel::paint(g);
-    }
-    
-    void resized() override
-    {
-        auto bounds = getLocalBounds();
-        
-        // Marges adaptatives pour éviter les problèmes aux petites tailles
-        int horizontalMargin = juce::jmin(4, bounds.getWidth() / 15);
-        int verticalMargin = juce::jmin(6, bounds.getHeight() / 8);
-        
-        auto area = bounds.reduced(horizontalMargin, verticalMargin);
-        flexBox.performLayout(area.toFloat());
-    }
-    
-    // Méthode pour calculer la taille préférée du composant
-    juce::Rectangle<int> getPreferredSize() const
-    {
-        // Calcul basé sur le nombre de boutons et leurs dimensions
-        int buttonWidth = 32;
-        int buttonSpacing = 4;
-        int padding = 16; // 8px de chaque côté
-        
-        int totalWidth = (3 * buttonWidth) + (2 * buttonSpacing) + padding; // 3 boutons, 2 espaces
-        int totalHeight = 32 + 12; // hauteur bouton + padding vertical
-        
-        return juce::Rectangle<int>(0, 0, totalWidth, totalHeight);
-    }
-    
-    // Accès aux boutons
-    StyledButton& getPlayButton() { return playButton; }
-    StyledButton& getPauseButton() { return pauseButton; }
-    StyledButton& getStopButton() { return stopButton; }
-    
-private:
-    StyledButton playButton;
-    StyledButton pauseButton;
-    StyledButton stopButton;
-    juce::FlexBox flexBox;
-    
-    JUCE_DECLARE_NON_COPYABLE_WITH_LEAK_DETECTOR(PlaybackActionArea)
-};
